+++ apps/@examples/checklist/src/components/index.ts
@@ -1,0 +1,3 @@
+export * from './button';
+export * from './radio';
+export * from './text';