+++ apps/@examples/checklist/src/components/text/index.tsx
@@ -1,0 +1,10 @@
+import { PropsWithChildren } from 'react';
+
+import { textClass } from './index.css';
+interface Props extends PropsWithChildren {
+  strikeThrough: boolean;
+}
+
+export function Text({ children, strikeThrough }: Props) {
+  return <span className={textClass({ strikeThrough })}>{children}</span>;
+}