+++ apps/@examples/embedded-contextual-education/src/components/tip-video.tsx
@@ -5,15 +5,18 @@ import {
   type ModalProps,
 } from '@chakra-ui/react';
 
-export interface TipVideoProps extends Omit<ModalProps, 'children'> {}
+export interface TipVideoProps extends Omit<ModalProps, 'children' | 'id'> {
+  id: string;
+}
 
 export function TipVideo(props: TipVideoProps) {
+  const { id, ...rest } = props;
   return (
-    <Modal size="3xl" {...props}>
+    <Modal size="3xl" {...rest}>
       <ModalOverlay />
       <ModalContent overflow="hidden">
         <iframe
-          src="https://www.youtube.com/embed/J---aiyznGQ?autoplay=true&modestbranding=true&rel=0"
+          src={`https://www.youtube.com/embed/${id}?autoplay=true&modestbranding=true&rel=0`}
           title="YouTube video player"
           allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share"
           allowFullScreen