+++ packages/@dopt/block-types/src/fields/fields.ts
@@ -1,8 +1,15 @@
 import { Static, Type } from '@sinclair/typebox';
 
+/**
+ * The supported types for a field's value: string, number, and boolean.
+ * This type also includes null for when the field's value is empty.
+ */
 export type FIELD_VALUE_UNION_TYPE = string | number | boolean | null;
-export type FIELD_TYPE_LITERALS = 'string' | 'number' | 'boolean';
-export const FIELD_TYPE_LITERALS: Record<string, FIELD_TYPE_LITERALS> = {
+/**
+ * The literal strings corresponding to field value types, "string", "number", and "boolean".
+ */
+export type FIELD_VALUE_LITERALS = 'string' | 'number' | 'boolean';
+export const FIELD_VALUE_LITERALS: Record<string, FIELD_VALUE_LITERALS> = {
   string: 'string',
   boolean: 'boolean',
   number: 'number',
@@ -21,7 +28,7 @@ export const StringField = Type.Intersect(
   [
     BaseField,
     Type.Object({
-      type: Type.Readonly(Type.Literal(FIELD_TYPE_LITERALS.string)),
+      type: Type.Readonly(Type.Literal(FIELD_VALUE_LITERALS.string)),
       value: Type.Readonly(Type.Union([Type.String(), Type.Null()])),
     }),
   ],
@@ -33,7 +40,7 @@ export const NumberField = Type.Intersect(
   [
     BaseField,
     Type.Object({
-      type: Type.Readonly(Type.Literal(FIELD_TYPE_LITERALS.number)),
+      type: Type.Readonly(Type.Literal(FIELD_VALUE_LITERALS.number)),
       value: Type.Readonly(Type.Union([Type.Number(), Type.Null()])),
     }),
   ],
@@ -45,7 +52,7 @@ export const BooleanField = Type.Intersect(
   [
     BaseField,
     Type.Object({
-      type: Type.Readonly(Type.Literal(FIELD_TYPE_LITERALS.boolean)),
+      type: Type.Readonly(Type.Literal(FIELD_VALUE_LITERALS.boolean)),
       value: Type.Readonly(Type.Union([Type.Boolean(), Type.Null()])),
     }),
   ],
@@ -56,6 +63,14 @@ export type BooleanField = Static<typeof BooleanField>;
 export const Field = Type.Union([StringField, NumberField, BooleanField], {
   $id: 'Field',
 });
+
+/**
+ * This type defines all the properties of a field.
+ * A field contains:
+ * - `sid`: a string, the identifier for the field
+ * - `type`: a string literal, one of "string", "number", or "boolean"
+ * - `value`: the value of the field, this must have type string, number, boolean, or `null` if the value is empty
+ */
 export type Field = Static<typeof Field>;
 
 export const Fields = Type.Array(Field);