+++ packages/@dopt/react/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 1.1.10
+
+### Patch Changes
+
+- ce2ac3f4: clean up type names and add JS / TS doc strings to create better typedocs for @dopt/react and @dopt/javascript packages
+- Updated dependencies [ce2ac3f4]
+  - @dopt/javascript-common@1.2.8
+  - @dopt/block-types@2.0.4
+  - @dopt/mercator@1.0.2
+  - @dopt/logger@0.0.3
+
 ## 1.1.9
 
 ### Patch Changes